Narrow remaining count/distance i16s to u8 in deduce/build/solve_brute
Continuation of the OptionValue cleanup: the cases left with i16 fell
into two camps, mechanical bounded counts and distance arithmetic that
was using `(a as i16 - b as i16).abs()` because the original sentinel
math went through i16.

- solve_brute.rs's count/remaining locals in CountAnswer / CountVowel /
  CountConsonant rule checking: u8 (bounded by MAX_N).
- deduce.rs's compute_letter_counts returns [u8; 5] arrays; `pos_contrib`
  becomes u8::from(...) with the comparison reordered (`> pos_contrib`)
  to avoid u8 underflow.
- deduce.rs's vowel+consonant cross-elim: explicit `is_num()` guards on
  both sides of the inner sum instead of folding NONE into -1; sums are
  now u8.
- deduce.rs's LetterDist forward / wrong / no-match / reverse: distance
  computed via u8::abs_diff, NONE treated as "can't match any actual
  distance" with explicit `s.is_num()` checks.
- build.rs's correct_option_value for LetterDist: u8::abs_diff.

Remaining 14 i16 uses are load-bearing: `Option::<i16>::deserialize` in
OptionValue (defensive against out-of-range JSON), `perturb_count`'s
signed offset arithmetic in build.rs (genuinely needs negative offsets),
and the test fixture comparison format (-1 = null convention in JSON).

gen-diff clean, 2029.json byte-identical, all suites pass.

+++ rust/src/solve_brute.rs
@@ -402,7 +402,7 @@ fn check_rule(
             if s.is_unused() {
                 return false;
             }
-            let opt_val = s.value() as i16;
+            let opt_val = s.value();
 
             let (range_start, range_end) = match *t {
                 QuestionType::CountAnswer { .. } => (0, n),
@@ -411,8 +411,8 @@ fn check_rule(
                 _ => unreachable!(),
             };
 
-            let mut count: i16 = 0;
-            let mut remaining: i16 = 0;
+            let mut count: u8 = 0;
+            let mut remaining: u8 = 0;
             for j in range_start..range_end {
                 if answers[j] == Some(answer) {
                     count += 1;
@@ -429,10 +429,10 @@ fn check_rule(
             if s.is_unused() {
                 return false;
             }
-            let opt_val = s.value() as i16;
+            let opt_val = s.value();
             let is_vowel = matches!(*t, QuestionType::CountVowel);
-            let mut count: i16 = 0;
-            let mut remaining: i16 = 0;
+            let mut count: u8 = 0;
+            let mut remaining: u8 = 0;
             for j in 0..n {
                 if answers[j].is_none() {
                     remaining += 1;
